Add functions getPlayersInput, placeMarker
Added a new variable activePlayer to switch the active player later.
The getPlayersInput-function returns the players input from two prompts
(row and column).
The placeMarker-function takes the picked cell and the active player as
argument to place the right marker on the right spot.
Added the playRound-function that calls getPlayersInput and placeMarker.