Add functions to make the game work
Added getBoard-function to return the board to the GameController.
Added getCellValue-function to return the values of a single cell to
the GameController.
Added getWinPatterns-function that returns 8 arrays (for the 8
win-patterns) from the board.
Added switchActivePlayer-function to switch the player after one turn.
Added checkWin-function that checks all win-patterns, if one of them
has 3 identical markers, the player with this marker wins, if the whole
board is filled with markers, but no pattern gets triggered, its a tie.
Edited playRound-function:
    1. get User-input
    2. check if the picked cell is available
    3. if available, place marker
          check if the player has won or if its a tie
          switch the active player
          play a new round
    4. if not available, replay the round without changing players