Add ui without functionality
Added a header and main-content part with some styles.
Added a board-div to the main-content-div.
Added a new function to create all cells of the board with js, the
"data-index-numbers"-attribute tells the row and column of the cell.
The board is created with a 3x3 grid-layout and a gap for the lines
between cells.