Add new ui-functions
Added redner-function in the gameboard to display all values from the
array in the ui.
Added eventlistener to gamecontroller to check with field was clicked,
it then calls the playRound-function.
Added fonts to the css-file.
Changed variable-name of the board in the gamecontroller from "board"
to "gameBoard" for better readability.
Deleted the "getCellValue"-function because it is useless and can be
bypassed.