Add form- and board-ui to html
Added all needed html-elements in the ".form-ui" so that the user can
input their name and start the game.
Added all needed html-elements in the ".board-ui" so that the user can
interact with the gameboard.
Removed the createCells-function because it is no longer needed to
create html-elements.
Added renderBoardUi- and renderFormUi-functions to switch between the
".form-ui"-html element and the ".board-ui"-html element.
Removed "render"-function as it is no longer needed.
Added EventListener to the ".form-ui" to get the player names and
switch to the ".board-ui".
Added new function "setPlayerNames" to set the names of the players.