Add new function "updateMessagePanel".
The new function is used to update the message-panel above the
gameboard, it was created to shorten up the render-function and for
better readability.
The render-function takes an object now, in this object are all
informations that are needed to render the right parts of the website.
Because of these changes, the updateMessagePanel-, showWinMessage-, and
showTieMessage-functions have been deleted.