Outsource eventlistener-functions
The functions, that get called by the event-listeners are now separate
for better readability.
The checkWin- and checkTie-functions are now named isWin and isTie,
because they always return a true or false, this makes more sense.
The playRound-function is now named startGameCycle because that is a
more acurate description.
The startGameCycle-function now prevents a further input from the user
if the game is already over.
Also, this function is now a little shorter, instead of calling the
render-function 3 times, it is only called once.