Add new functions "restartGame" and "startNewGame"
Added eventlistener, it calls the new function "restartGame", this
function resets the board and the active player and renders the board
and messagePanel.
Added eventlistener, it calls the new function "startNewGame", this
function calls "restartGame" and then renders the form to start a new
game with new names.
The "render" function has been changed because an programming-mistake,
the “updateBoard”-function was called when rendering the messagePanel,
now the "updateBoard"-function gets called in the right place
(when rendering the board).