Add logic to display the markers on the board.
Changed "updateBoard"-function to display the x- and o-svg's on the
board and delete them if a new game starts.
Changed "updateMessagePanel"-function to display the x- and o-svg's on
it.
Added svg-images in css.
Changed some color values (more color-variables).
Changed background-image because it looks better.
Added marker-svg's to the input-fields, so that the players know who
plays with x and who plays with o.
Changed selector "message-panel" to "game-info" because inside of
"game-info" is now a message-panel and a marker-panel.